+++ R/tm_t_summary.R
@@ -6,7 +6,8 @@
 #' @param summarize_vars \code{\link[teal]{choices_selected}} object with all available choices and preselected option
 #'   for variable names that can be used for summary
 #'
-#' @importFrom rtables as_html
+#' @importFrom dplyr select
+#' @importFrom rtables as_html var_relabel
 #'
 #' @export
 #'
@@ -15,19 +16,17 @@
 #' @examples
 #' library(random.cdisc.data)
 #'
-#' ASL <- radsl(seed = 1)
-#' keys(ASL) <- c("STUDYID", "USUBJID")
+#' ADSL <- radsl(cached = TRUE)
 #'
 #' app <- init(
 #'   data = cdisc_data(
-#'     ASL = ASL,
-#'     code = 'ASL <- radsl(seed = 1)
-#'             keys(ASL) <- c("STUDYID", "USUBJID")',
+#'     cdisc_dataset("ADSL", ADSL),
+#'     code = 'ADSL <- radsl(cached = TRUE)',
 #'     check = FALSE),
 #'   modules = root_modules(
 #'     tm_t_summary(
 #'       label = "Demographic Table",
-#'       dataname = "ASL",
+#'       dataname = "ADSL",
 #'       arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
 #'       summarize_vars = choices_selected(c("SEX", "RACE", "BMRKR2"), c("SEX", "RACE"))
 #'     )
@@ -75,6 +74,7 @@ ui_t_summary <- function(id, ...) {
                           a$arm_var$choices,
                           a$arm_var$selected,
                           multiple = FALSE),
+      checkboxInput(ns("add_total"), "Add All Patients column", value = TRUE),
       optionalSelectInput(ns("summarize_vars"),
                           "Summarize Variables",
                           a$summarize_vars$choices,
@@ -89,44 +89,62 @@ ui_t_summary <- function(id, ...) {
 }
 
 srv_t_summary <- function(input, output, session, datasets, dataname) {
+
   init_chunks()
 
-  table_call <- reactive({
+  output$table <- renderUI({
     anl_f <- datasets$get_data(dataname, reactive = TRUE, filtered = TRUE)
 
     arm_var <- input$arm_var
+    add_total <- input$add_total
     summarize_vars <- input$summarize_vars
 
-    validate_has_data(anl_f, min_nrow = 3)
+    validate(need(is.logical(add_total), "add total is not logical"))
     validate(need(!is.null(summarize_vars), "please select 'summarize variables'"))
     validate(need(all(summarize_vars %in% names(anl_f)), "not all variables available"))
-    validate(need(anl_f[[arm_var]], "Arm variable does not exist"))
-    validate(need(!("" %in% anl_f[[arm_var]]), "arm values can not contain empty strings ''"))
+    validate(need(!is.null(arm_var), "please select 'arm variable'"))
+    validate(need(arm_var %in% names(anl_f), "arm variable does not exist"))
+    validate_has_data(anl_f, min_nrow = 3)
 
     data_name <- paste0(dataname, "_FILTERED")
     assign(data_name, anl_f)
 
     chunks_reset(envir = environment())
 
-    table_chunk_expr <- bquote({
+    chunks_push(
+      call(
+        "<-",
+        as.name("ANL"),
+        Reduce(
+          function(x, y) call("%>%", x, y),
+          c(
+            as.name(data_name),
+            as.call(c(
+              list(quote(dplyr::select)),
+              lapply(c(summarize_vars, arm_var), as.name)
+            )),
+            as.call(append(
+              quote(rtables::var_relabel),
+              datasets$get_data_attr(dataname = "ADSL", "labels")$column_labels[summarize_vars]
+            ))
+          )
+        )
+      )
+    )
+
+    total <- if (add_total) "All Patients" else NULL
+
+    chunks_push(bquote({
       tbl <- t_summary(
-        x = .(as.name(data_name))[, .(summarize_vars), drop = FALSE],
-        col_by = as.factor(.(as.name(data_name))[[.(arm_var)]]),
-        total = "All Patients",
+        x = .(as.name("ANL"))[, .(summarize_vars), drop = FALSE],
+        col_by = as.factor(.(as.name("ANL"))[[.(arm_var)]]),
+        total = .(total),
         useNA = "ifany"
       )
       tbl
-    })
-    chunks_push(expression = table_chunk_expr, id = "tm_t_summary_tbl")
+    }))
 
-    return(invisible(NULL))
-  })
-
-  output$table <- renderUI({
-    table_call()
-
-    chunks_eval()
-    chunks_validate_all("tbl", "rtable", "Evaluation with tern t_tte failed.")
+    chunks_safe_eval()
 
     tbl <- chunks_get_var("tbl")
     as_html(tbl)