+++ R/tm_t_tte.R
@@ -12,7 +12,7 @@
 #' @param arm_var \code{\link[teal]{choices_selected}} object with all available choices and preselected option
 #' for variable names that can be used as \code{arm_var}
 #' @param arm_ref_comp (\code{\link[teal]{choices_selected}}) optional, if specified it must be a named list with each
-#'   element corresponding to an arm variable in \code{ASL} and the element must
+#'   element corresponding to an arm variable in \code{ADSL} and the element must
 #'   be another list with the elements named \code{ref} and \code{comp} that the
 #'   defined the default reference and comparison arms when the arm variable is
 #'   changed.
@@ -37,7 +37,7 @@
 #'  filtering for \code{paramcd} one observation per patient is expected
 #' }
 #'
-#' The arm variables, stratification variables and taken from the \code{ASL}
+#' The arm variables, stratification variables and taken from the \code{ADSL}
 #' data.
 #'
 #'
@@ -50,23 +50,21 @@
 #' @examples
 #' library(random.cdisc.data)
 #'
-#' ASL <- radsl(seed = 1)
-#' ATE <- radtte(ASL, seed = 1)
+#' ADSL <- radsl(cached = TRUE)
+#' ADTTE <- radtte(ADSL, cached = TRUE)
 #'
-#' keys(ASL) <- keys(ATE) <- c("USUBJID", "STUDYID")
 #'
 #' app <- init(
-#'     data = cdisc_data(ASL = ASL, ATE = ATE,
-#'         code = "ASL <- radsl(seed = 1)
-#'                 ATE <- radtte(ASL, seed = 1)
-#'                 keys(ASL) <- keys(ATE) <- c('USUBJID', 'STUDYID')",
+#'     data = cdisc_data(cdisc_dataset("ADSL", ADSL), cdisc_dataset("ADTTE", ADTTE),
+#'         code = "ADSL <- radsl(cached = TRUE)
+#'                 ADTTE <- radtte(ADSL, cached = TRUE)",
 #'         check = FALSE),
 #'     modules = root_modules(
 #'         tm_t_tte(
 #'             label = "Time To Event Table",
-#'             dataname = 'ATE',
+#'             dataname = 'ADTTE',
 #'             arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
-#'             paramcd = choices_selected(unique(ATE$PARAMCD), "OS"),
+#'             paramcd = choices_selected(unique(ADTTE$PARAMCD), "OS"),
 #'             strata_var = choices_selected(c("SEX", "BMRKR2"), "SEX"),
 #'             time_points = choices_selected(c(6, 8), 6),
 #'             time_unit = "month",
@@ -77,9 +75,7 @@
 #'
 #'
 #' \dontrun{
-#'  \donttest{
-#'    shinyApp(app$ui, app$server)
-#'   }
+#'  shinyApp(app$ui, app$server)
 #' }
 #'
 #' ## Define default reference & comparison arms based on
@@ -88,11 +84,12 @@
 #' library(dplyr)
 #' library(magrittr)
 #'
-#' ASL <- mutate(radsl(seed = 1),
+#' ADSL <- mutate(radsl(seed = 1),
 #'   ARM1 = sample(c("DUMMY A", "DUMMY B"),
 #'   n(), TRUE))
-#' ATE <- radtte(ASL, seed = 1)
-#' keys(ASL) <- keys(ATE) <- c("USUBJID", "STUDYID")
+#' attr(ADSL$ARM1, "label") <- "dummy"
+#' ADTTE <- radtte(ADSL, seed = 1)
+#'
 #'
 #' arm_ref_comp = list(
 #'   ACTARMCD = list(
@@ -105,19 +102,18 @@
 #'   )
 #' )
 #' app <- init(
-#'     data = cdisc_data(ASL = ASL, ATE = ATE,
-#'         code = "ASL <- radsl(seed = 1) %>%
+#'     data = cdisc_data(cdisc_dataset("ADSL", ADSL), cdisc_dataset("ADTTE", ADTTE),
+#'         code = "ADSL <- radsl(seed = 1) %>%
 #'                 mutate(., ARM1 = sample(c('DUMMY A', 'DUMMY B'), n(), TRUE))
-#'                 ATE <- radtte(ASL, seed = 1)
-#'                 keys(ASL) <- keys(ATE) <- c('USUBJID', 'STUDYID')",
+#'                 ADTTE <- radtte(ADSL, seed = 1)",
 #'         check = FALSE),
 #'     modules = root_modules(
 #'         tm_t_tte(
 #'          label = "Time To Event Table",
-#'          dataname = 'ATE',
+#'          dataname = 'ADTTE',
 #'          arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
 #'          arm_ref_comp = arm_ref_comp,
-#'          paramcd = choices_selected(unique(ATE$PARAMCD), "OS"),
+#'          paramcd = choices_selected(unique(ADTTE$PARAMCD), "OS"),
 #'          strata_var = choices_selected(c("SEX", "MLIVER"), "SEX"),
 #'          time_points = choices_selected(c(6, 8), 6),
 #'          time_unit = "months",
@@ -207,7 +203,7 @@ ui_t_tte <- function(id, ...) {
                           a$strata_var$choices,
                           a$strata_var$selected,
                           multiple = TRUE,
-                          label_help = helpText("from ", tags$code("ASL"))),
+                          label_help = helpText("from ", tags$code("ADSL"))),
       optionalSelectInput(ns("time_points"),
                           "Time Points",
                           a$time_points$choices,
@@ -242,7 +238,7 @@ srv_t_tte <- function(input,
   arm_ref_comp_observer(
     session, input,
     id_ref = "ref_arm", id_comp = "comp_arm", id_arm_var = "arm_var",    # from UI
-    asl = datasets$get_data("ASL", filtered = FALSE, reactive = FALSE),
+    adsl = datasets$get_data("ADSL", filtered = FALSE, reactive = FALSE),
     arm_ref_comp = arm_ref_comp,
     module = "tm_t_tte"
   )
@@ -251,7 +247,7 @@ srv_t_tte <- function(input,
   table_reactive <- reactive({
     # resolve all reactive expressions
     # nolint start
-    ASL_FILTERED <- datasets$get_data("ASL", reactive = TRUE, filtered = TRUE)
+    ADSL_FILTERED <- datasets$get_data("ADSL", reactive = TRUE, filtered = TRUE)
     ANL_FILTERED <- datasets$get_data(dataname, reactive = TRUE, filtered = TRUE)
     # nolint end
 
@@ -275,8 +271,8 @@ srv_t_tte <- function(input,
 
     # validate your input values
     validate_standard_inputs(
-      asl = ASL_FILTERED,
-      aslvars = c("USUBJID", "STUDYID", arm_var, strata_var),
+      adsl = ADSL_FILTERED,
+      adslvars = c("USUBJID", "STUDYID", arm_var, strata_var),
       anl = ANL_FILTERED,
       anlvars = c("USUBJID", "STUDYID",  "PARAMCD", "AVAL", "CNSR", event_desc_var),
       arm_var = arm_var,
@@ -293,37 +289,39 @@ srv_t_tte <- function(input,
 
     chunks_reset(envir = environment())
 
-    asl_vars <- unique(c("USUBJID", "STUDYID", arm_var, strata_var))
+    adsl_vars <- unique(c("USUBJID", "STUDYID", arm_var, strata_var))
     anl_vars <- unique(c("USUBJID", "STUDYID", "AVAL", "CNSR", event_desc_var))
 
     ## Now comes the analysis code
-    chunks_push(expression = bquote(ref_arm <- .(ref_arm)))
-    chunks_push(expression = bquote(comp_arm <- .(comp_arm)))
-    chunks_push(expression = bquote(strata_var <- .(strata_var)))
-    chunks_push(expression = bquote(combine_comp_arms <- .(combine_comp_arms)))
+    chunks_push(bquote(ref_arm <- .(ref_arm)))
+    chunks_push(bquote(comp_arm <- .(comp_arm)))
+    chunks_push(bquote(strata_var <- .(strata_var)))
+    chunks_push(bquote(combine_comp_arms <- .(combine_comp_arms)))
 
-    chunks_push(expression = bquote(asl_p <- subset(ASL_FILTERED, .(as.name(arm_var)) %in% c(ref_arm, comp_arm))))
-    chunks_push(expression = bquote(anl_endpoint <- subset(.(as.name(anl_name)), PARAMCD == .(paramcd))))
+    chunks_push(bquote(adsl_p <- subset(ADSL_FILTERED, .(as.name(arm_var)) %in% c(ref_arm, comp_arm))))
+    chunks_push(bquote(anl_endpoint <- subset(.(as.name(anl_name)), PARAMCD == .(paramcd))))
 
-    chunks_push(expression = bquote(anl <- merge(
-      x = asl_p[, .(asl_vars)],
-      y = anl_endpoint[, .(anl_vars)],
-      all.x = FALSE, all.y = FALSE,
-      by = c("USUBJID", "STUDYID")
-    )))
+    chunks_push(bquote({
+      anl <- merge(
+        x = adsl_p[, .(adsl_vars)],
+        y = anl_endpoint[, .(anl_vars)],
+        all.x = FALSE, all.y = FALSE,
+        by = c("USUBJID", "STUDYID")
+      )
+    }))
 
-    chunks_push(expression = bquote(arm <- relevel(as.factor(anl[[.(arm_var)]]), ref_arm[1])))
-    chunks_push(expression = bquote(arm <- combine_levels(arm, ref_arm)))
+    chunks_push(bquote(arm <- relevel(as.factor(anl[[.(arm_var)]]), ref_arm[1])))
+    chunks_push(bquote(arm <- combine_levels(arm, ref_arm)))
     if (combine_comp_arms) {
-      chunks_push(expression = bquote(arm <- combine_levels(arm, comp_arm)))
+      chunks_push(bquote(arm <- combine_levels(arm, comp_arm)))
     }
-    chunks_push(expression = bquote(anl[[.(arm_var)]] <- droplevels(arm)))
+    chunks_push(bquote(anl[[.(arm_var)]] <- droplevels(arm)))
 
-    chunks_eval()
+    chunks_safe_eval()
 
     validate(need(nrow(chunks_get_var("anl")) > 15, "need at least 15 data points"))
 
-    table_expr <- bquote({
+    chunks_push(bquote({
       tbl <- t_tte(
         formula = .(as.formula(
           paste0(
@@ -332,23 +330,22 @@ srv_t_tte <- function(input,
               ""
             } else {
               paste0(" + strata(", paste(strata_var, collapse = ", "), ")")
-            }))),
+            }
+          )
+        )),
         data = anl,
         event_descr = if (is.null(.(event_desc_var))) NULL else as.factor(anl[[.(event_desc_var)]]),
         time_points = .(time_points),
         time_unit = .(time_unit)
       )
       tbl
-    })
-
-    chunks_push(expression = table_expr, id = "final_table")
+    }))
   })
 
   output$tte_table <- renderUI({
     table_reactive()
 
-    chunks_eval()
-    chunks_validate_all("tbl", "rtable", "Evaluation with tern t_tte failed.")
+    chunks_safe_eval()
 
     tbl <- chunks_get_var("tbl")
     as_html(tbl)